NY Trip! Successful lighting, new trippy shader
Long drive. What else was I gonna do? Inspired by Picasso.

+++ CodeIndigo/Indigo/World.h
@@ -33,7 +33,7 @@ public:
 	// Compiles and puts in place custom Vertex or Fragment Shaders
 	void Shader(const char * vertex, const char * fragment);
 	// Get the location of a variable in a shader
-	unsigned short Shader_Location(const char * name, const bool uniform = false);
+	int Shader_Location(const char * name, const bool uniform = false);
 
 	/*
 	Add a skybox, a backdrop to the world. Accepts a filename of a texture.
